requirements.txt
alabaster==0.7.6
amqp==1.4.9
analytics-python==1.2.3
anyjson==0.3.3
Babel==2.1.1
beautifulsoup4==4.2.1
billiard==3.3.0.23
boto==2.38.0
boto3==1.3.1
botocore==1.4.28
braintree==3.25.0
celery==3.1.23
cffi==1.6.0
Cheetah==2.4.4
configobj==4.7.2
cryptography==1.4
decorator==3.4.0
diff-match-patch==20121119
Django==1.8.3
django-allauth==0.23.0
django-authtools==1.2.0
django-braces==1.8.1
django-cors-headers==1.1.0
django-dajax==0.9.2
django-dajaxice==0.7
django-debug-toolbar==1.3.2
django-djconfig==0.4.0
django-extensions==1.5.5
django-filter==0.10.0
django-frontend==1.5.0
django-frontend-skeleton==3.0.0
django-haystack==2.4.0
django-import-export==0.2.7
django-infinite-scroll-pagination==0.1.3
django-oauth-toolkit==0.9.0
django-rest-auth==0.5.0
django-rest-framework-social-oauth2==1.0.2
django-spirit==0.4.2
django-sslserver==0.15
django-storages==1.1.8
django-vanilla-views==1.0.3
djangorestframework==3.1.3
docutils==0.11
enum34==1.1.6
futures==3.0.3
html5lib==0.999
idna==2.1
ipaddress==1.0.16
ipython==1.2.1
Jinja2==2.7.2
jmespath==0.9.0
joblib==0.7.1
jsonfield==1.0.3
jsonpatch==1.3
jsonpointer==1.0
kombu==3.0.35
lxml==3.4.4
Markdown==2.4
MarkupSafe==0.18
matplotlib==1.3.1
mem-top==0.1.2
mistune==0.7.3
MySQL-python==1.2.5
ndg-httpsclient==0.4.1
nltk==3.0.2
nose==1.3.1
numpy
numexpr
oauth==1.0.1
oauthlib==1.0.1
openpyxl==1.7.0
PAM
pandas
patsy==0.2.1
pexpect==3.1
Pillow==2.9.0
plotly
polyline==1.3.1
prettytable==0.7.2
pyasn1==0.1.9
pycparser==2.14
pycurl==7.19.3
Pygments==2.0.2
PyJWT==1.4.0
pyparsing==2.0.1
pyserial==2.6
python-dateutil==2.5.3
python-Levenshtein==0.12.0
python-openid==2.2.5
python-social-auth==0.2.12
pytz
PyYAML==3.10
pyzmq==14.0.1
redis==2.10.5
requests==2.7.0
requests-oauthlib==0.4.2
roman==2.0.0
s3==3.0.0
scikit-learn==0.16.1
scipy
simplegeneric==0.8.1
simplejson==3.3.1
six==1.9.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
sqlparse==0.1.16
ssh-import-id==3.21
statsmodels
sympy==0.7.4.1
tables
tablib==0.10.0
textile==2.1.5
tornado
unicode-slugify==0.1.3
Unidecode==0.4.18
urllib3==1.7.1
Whoosh==2.7.0
xlrd==0.9.2
xlwt==0.7.5
xmltodict==0.9.2
zope.interface==4.0.5